Finished speed trace view

+++ pitwall-ios/pitwall-ios/Telemetry.swift
@@ -27,21 +27,21 @@ func integrate(pt1: [Double], pt2: [Double]) -> Double {
     return triangleArea + rectangularArea
 }
 
-func addDistance(CarData: [(value: String, timestamp: Double)]) -> [(speed: Double, distance: Double)] {
-    var returnArr = [(speed: Double, distance: Double)]()
+func addDistance(CarData: [(value: String, timestamp: Double)]) -> [(speed: Int, distance: Double)] {
+    var returnArr = [(speed: Int, distance: Double)]()
     
     for (i, item) in CarData.enumerated() {
         let timestamp = item.timestamp
         let speed = Double(item.value.components(separatedBy: ",")[1]) ?? 0
         
         if i == 0 { // Distance travelled is 0 on first sample
-            returnArr.append((speed: speed, distance: 0))
+            returnArr.append((speed: Int(speed), distance: 0))
         } else {
             let prev = CarData[i-1]
             let timestampPrev = prev.timestamp
             let speedPrev = Double(prev.value.components(separatedBy: ",")[1]) ?? 0
             let distance = integrate(pt1: [timestampPrev, speedPrev], pt2: [timestamp, speed]) + (returnArr.last?.distance ?? 0)
-            returnArr.append((speed: speed, distance: distance))
+            returnArr.append((speed: Int(speed), distance: distance))
         }
     }
     return returnArr
